Introduce functional tests
