+++ pipelines/training/autorag/documents_rag_optimization_pipeline/tests/.env.example
@@ -17,10 +17,10 @@ LLAMA_STACK_SECRET_NAME=
 LLAMA_STACK_VECTOR_IO_PROVIDER_ID=
 
 # S3-compatible storage for artifact checks (optional; same as pipeline backend storage)
-AWS_S3_ENDPOINT=
-AWS_ACCESS_KEY_ID=
-AWS_SECRET_ACCESS_KEY=
-AWS_DEFAULT_REGION=us-east-1
+ARTIFACTS_AWS_ACCESS_KEY_ID=
+ARTIFACTS_AWS_SECRET_ACCESS_KEY=
+ARTIFACTS_AWS_S3_ENDPOINT=
+ARTIFACTS_AWS_DEFAULT_REGION="us-east-1"
 RHOAI_TEST_ARTIFACTS_BUCKET=
 
 # Optional: timeout in seconds for pipeline run (default 3600)
@@ -29,13 +29,36 @@ RHOAI_TEST_ARTIFACTS_BUCKET=
 # Optional: skip TLS verify for self-signed cluster certs (true/false)
 # KFP_VERIFY_SSL=false
 
+# Kubernetes API for pod log fetching on failure (optional overrides)
+# K8S_API_URL overrides the entire derived URL; K8S_API_PORT overrides just the port
+# Default port: 443 for ROSA clusters, 6443 for standard OCP
+# K8S_API_URL=https://api.my-cluster.example.com:443
+# K8S_API_PORT=443
+
 # Functional test: milvus provider IDs (one per milvus mode)
 LLAMA_STACK_VECTOR_IO_PROVIDER_ID_MILVUS_LITE=
-LLAMA_STACK_VECTOR_IO_PROVIDER_ID_MILVUS_STANDALONE=
+LLAMA_STACK_VECTOR_IO_PROVIDER_ID_MILVUS_REMOTE=
 
 # Functional test: constrained model lists (JSON arrays, e.g. '["model-a", "model-b"]')
-# DOCRAG_EMBEDDINGS_MODELS=
-# DOCRAG_GENERATION_MODELS=
+# FUNC_TEST_EMBEDDINGS_MODELS=
+# FUNC_TEST_GENERATION_MODELS=
 
 # Functional test: filter test configs by tags (comma-separated, e.g. smoke,milvus-lite)
 # RHOAI_TEST_CONFIG_TAGS=
+
+# =============================================================================
+# Env varialbes to test notebooks (can be extraced from kubernetes secrets as well)
+# Note: these variables need to have the same names as one would expect in production
+# =============================================================================
+
+# Llama Stack env variables
+LLAMA_STACK_CLIENT_BASE_URL=
+LLAMA_STACK_CLIENT_API_KEY=
+
+
+# AWS env varaibles for the indexing notebooks execution
+AWS_ACCESS_KEY_ID=
+AWS_SECRET_ACCESS_KEY=
+AWS_S3_ENDPOINT=
+AWS_DEFAULT_REGION=us-east-1
+AWS_S3_BUCKET=
